Refactoring and adding UPN to request
