+++ server/plugin.go
@@ -4,35 +4,21 @@
 package main
 
 import (
-	"encoding/json"
-	"fmt"
 	"io/ioutil"
 	"path/filepath"
 	"sync"
 
-	"github.com/mattermost/mattermost-plugin-msteams-meetings/server/remote"
+	"github.com/mattermost/mattermost-plugin-msteams-meetings/server/store"
 	"github.com/mattermost/mattermost-server/v5/model"
 	"github.com/mattermost/mattermost-server/v5/plugin"
 	"github.com/pkg/errors"
-	msgraph "github.com/yaegashi/msgraph.go/beta"
-	"golang.org/x/oauth2"
-	"golang.org/x/oauth2/microsoft"
 )
 
 const (
-	postMeetingKey = "post_meeting_"
-
 	botUserName    = "mstmeetings"
 	botDisplayName = "MS Teams Meetings"
 	botDescription = "Created by the MS Teams Meetings plugin."
-
-	tokenKey           = "token_"
-	tokenKeyByRemoteID = "tbyrid_"
-
-	stateLength = 3
 )
-
-var oAuthMessage string = "[Click here to link your Microsoft account.](%s/plugins/" + manifest.Id + "/oauth2/connect?channelID=%s)"
 
 // Plugin defines the plugin struct
 type Plugin struct {
@@ -47,6 +33,8 @@ type Plugin struct {
 	// configuration is the active plugin configuration. Consult getConfiguration and
 	// setConfiguration for usage.
 	configuration *configuration
+
+	store store.Store
 }
 
 // OnActivate checks if the configurations is valid and ensures the bot account exists
@@ -88,178 +76,9 @@ func (p *Plugin) OnActivate() error {
 		return errors.Wrap(appErr, "couldn't set profile image")
 	}
 
-	return nil
-}
-
-func (p *Plugin) getSiteURL() (string, error) {
-	siteURLRef := p.API.GetConfig().ServiceSettings.SiteURL
-	if siteURLRef == nil || *siteURLRef == "" {
-		return "", errors.New("error fetching siteUrl")
+	p.store = store.Store{
+		API: p.API,
 	}
 
-	return *siteURLRef, nil
-}
-
-func (p *Plugin) getOAuthConfig() (*oauth2.Config, error) {
-	config := p.getConfiguration()
-
-	clientID := config.OAuth2ClientID
-	clientSecret := config.OAuth2ClientSecret
-	clientAuthority := config.OAuth2Authority
-
-	siteURL, err := p.getSiteURL()
-	if err != nil {
-		return nil, err
-	}
-
-	redirectURL := fmt.Sprintf("%s/plugins/%s/oauth2/complete", siteURL, manifest.Id)
-
-	return &oauth2.Config{
-		ClientID:     clientID,
-		ClientSecret: clientSecret,
-		RedirectURL:  redirectURL,
-		Scopes: []string{
-			"offline_access",
-			"OnlineMeetings.ReadWrite",
-		},
-		Endpoint: microsoft.AzureADEndpoint(clientAuthority),
-	}, nil
-}
-
-// UserInfo defines the information we store from each user
-type UserInfo struct {
-	Email      string
-	OAuthToken *oauth2.Token
-	// Mattermost userID
-	UserID string
-	// Remote userID
-	RemoteID string
-}
-
-type authError struct {
-	Message string `json:"message"`
-	Err     error  `json:"err"`
-}
-
-func (ae *authError) Error() string {
-	errorString, _ := json.Marshal(ae)
-	return string(errorString)
-}
-
-func (p *Plugin) storeUserInfo(info *UserInfo) error {
-	jsonInfo, err := json.Marshal(info)
-	if err != nil {
-		return err
-	}
-
-	if err := p.API.KVSet(tokenKey+info.UserID, jsonInfo); err != nil {
-		return err
-	}
-
-	if err := p.API.KVSet(tokenKeyByRemoteID+info.RemoteID, jsonInfo); err != nil {
-		return err
-	}
-
-	return nil
-}
-
-func (p *Plugin) getUserInfo(userID string) (*UserInfo, error) {
-	var userInfo UserInfo
-
-	infoBytes, appErr := p.API.KVGet(tokenKey + userID)
-	if appErr != nil || infoBytes == nil {
-		return nil, errors.New("must connect user account to Microsoft first")
-	}
-
-	err := json.Unmarshal(infoBytes, &userInfo)
-	if err != nil {
-		return nil, errors.New("unable to parse token")
-	}
-
-	return &userInfo, nil
-}
-
-func (p *Plugin) authenticateAndFetchUser(userID, userEmail, channelID string) (*msgraph.User, *authError) {
-	var user *msgraph.User
-	var err error
-
-	siteURL, err := p.getSiteURL()
-	if err != nil {
-		p.API.LogError("authenticateAndFetchUser, cannot get site URL", "error", err.Error())
-		return nil, &authError{Message: "Cannot get Site URL. Contact your sys admin.", Err: err}
-	}
-
-	userInfo, apiErr := p.getUserInfo(userID)
-	oauthMsg := fmt.Sprintf(
-		oAuthMessage,
-		siteURL, channelID)
-
-	if apiErr != nil || userInfo == nil {
-		return nil, &authError{Message: oauthMsg, Err: apiErr}
-	}
-	user, err = p.getUserWithToken(userInfo.OAuthToken)
-	if err != nil || user == nil {
-		return nil, &authError{Message: oauthMsg, Err: apiErr}
-	}
-
-	return user, nil
-}
-
-func (p *Plugin) disconnect(userID string) error {
-	rawInfo, appErr := p.API.KVGet(tokenKey + userID)
-	if appErr != nil {
-		return appErr
-	}
-
-	var info UserInfo
-	err := json.Unmarshal(rawInfo, &info)
-	if err != nil {
-		return err
-	}
-
-	errByMattermostID := p.API.KVDelete(tokenKey + userID)
-	errByRemoteID := p.API.KVDelete(tokenKeyByRemoteID + info.RemoteID)
-	if errByMattermostID != nil {
-		return errByMattermostID
-	}
-	if errByRemoteID != nil {
-		return errByRemoteID
-	}
-	return nil
-}
-
-func (p *Plugin) getUserWithToken(token *oauth2.Token) (*msgraph.User, error) {
-	conf, err := p.getOAuthConfig()
-	if err != nil {
-		return nil, err
-	}
-
-	client := remote.NewClient(conf, token, p.API)
-
-	user, err := client.GetMe()
-	if err != nil {
-		return nil, err
-	}
-
-	return user, nil
-}
-
-func (p *Plugin) dm(userID string, message string) error {
-	channel, err := p.API.GetDirectChannel(userID, p.botUserID)
-	if err != nil {
-		p.API.LogInfo("couldn't get bot's DM channel", "user_id", userID, "bot_id", p.botUserID, "error", err.Error())
-		return err
-	}
-
-	post := &model.Post{
-		Message:   message,
-		ChannelId: channel.Id,
-		UserId:    p.botUserID,
-	}
-
-	_, err = p.API.CreatePost(post)
-	if err != nil {
-		return err
-	}
 	return nil
 }
