+++ server/remote/meeting.go
@@ -8,24 +8,25 @@ import (
 	"net/http"
 	"time"
 
+	"github.com/mattermost/mattermost-plugin-msteams-meetings/server/store"
 	"github.com/pkg/errors"
 	msgraph "github.com/yaegashi/msgraph.go/beta"
 )
 
-func (c *Client) CreateMeeting(userID string, attendeesIDs ...string) (*msgraph.OnlineMeeting, error) {
+func (c *Client) CreateMeeting(creator *store.UserInfo, attendeesIDs ...*store.UserInfo) (*msgraph.OnlineMeeting, error) {
 	ctx := context.Background()
 	start := time.Now()
 	end := start.Add(1 * time.Hour)
 	subject := "Mattermost Meeting"
 	attendees := []msgraph.MeetingParticipantInfo{}
-	for _, attendeeID := range attendeesIDs {
-		id := attendeeID
+	for _, attendee := range attendeesIDs {
 		attendees = append(attendees, msgraph.MeetingParticipantInfo{
 			Identity: &msgraph.IdentitySet{
 				User: &msgraph.Identity{
-					ID: &id,
+					ID: &attendee.RemoteID,
 				},
 			},
+			Upn: &attendee.UPN,
 		})
 	}
 
@@ -37,16 +38,17 @@ func (c *Client) CreateMeeting(userID string, attendeesIDs ...string) (*msgraph.
 			Organizer: &msgraph.MeetingParticipantInfo{
 				Identity: &msgraph.IdentitySet{
 					User: &msgraph.Identity{
-						ID: &userID,
+						ID: &creator.RemoteID,
 					},
 				},
+				Upn: &creator.UPN,
 			},
 			Attendees: attendees,
 		},
 	}
 	out := msgraph.OnlineMeeting{}
 
-	err := c.builder.Users().ID(userID).OnlineMeetings().Request().JSONRequest(ctx, http.MethodPost, "", &in, &out)
+	err := c.builder.Users().ID(creator.RemoteID).OnlineMeetings().Request().JSONRequest(ctx, http.MethodPost, "", &in, &out)
 	if err != nil {
 		return nil, errors.Wrap(err, "cannot create meeting")
 	}
